tous les controllers fonctionnent, avec gérance des droits admin ou non

+++ backend/controllers/commentController.js
@@ -51,7 +51,29 @@ exports.createComment = async (req, res)=>{
 
 exports.modifyComment = async (req, res)=>{
     try {
+        const sqlAdmin = `SELECT * FROM comments WHERE comment_id = ${req.params.id}`;
+        database.query(sqlAdmin, (error, result)=>{
+            if (error) res.status(400).json("Erreur affichage comm à supprimer " + error);
+            const userOfComment = result[0].user_id;
 
+            if(userOfComment == req.auth.userId || req.auth.admin == 1){
+                // On créé le contenu qui va changer
+            const updateComment = {
+                comment_id: req.params.id,
+                comment_text: req.body.text,
+                user_id: req.auth.userId,
+                comment_createdat: new Date() // Important car la date resterait l'ancienne
+            };
+            // Requete isolée et lancée
+            const sql = "UPDATE comments SET ?";
+            database.query(sql, updateComment, (error, result)=>{
+                if(error) res.status(400).json(error);
+                res.status(200).json("Commentaire mis à jour !");
+        });
+            }else{
+                res.status(400).json("Action non autorisée");
+            } 
+        })  
     } catch (error) {
         res.status(400).json("Erreur modification comment " + error);
     };
@@ -59,13 +81,27 @@ exports.modifyComment = async (req, res)=>{
 
 exports.deleteComment = async (req, res)=>{
     try {
-        const sql = "DELETE FROM comments WHERE comment_id = ?";
-        const commentId = req.params.id;
-        database.query(sql, commentId, (error, result)=>{
-            if(!result) throw error;
-            res.status(200).json("Comment supprimé");
-        });
+        const sqlAdmin = `SELECT * FROM comments WHERE comment_id = ${req.params.id}`;
+        database.query(sqlAdmin, (error, result)=>{
+            if (error) res.status(400).json("Erreur affichage comm à supprimer " + error);
+            const userOfComment = result[0].user_id;
+
+                if(userOfComment == req.auth.userId || req.auth.admin == 1){
+                    // Basique, bien préciser le comment id
+                    const sql = "DELETE FROM comments WHERE comment_id = ?";
+                    const commentId = req.params.id;
+                    database.query(sql, commentId, (error, result)=>{
+                        if(!result) throw error;
+                        res.status(200).json("Comment supprimé");
+                    });
+                }else{
+                    res.status(400).json("Action non autorisée");
+                } 
+        });  
     } catch (error) {
         res.status(400).json("Erreur suppression comment " + error);
     };
 };
+
+
+
